requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
pymongo
scikit-learn
numpy